requirements (2).txt
streamlit
tempfile
model
langdetect
tesseract
langid
collections 
counter
PIL
OCR